Split monolithic main.py into modular package
